+++ source/renderer/app/api/network/types.js
@@ -29,7 +29,7 @@ export type EpochLength = {
   unit: string,
 };
 
-export type EpochStability = {
+export type SecurityParameter = {
   quantity: number,
   unit: string,
 };
@@ -102,7 +102,7 @@ export type GetNetworkParametersResponse = {
   blockchainStartTime: number,
   slotLength: SlotLength,
   epochLength: EpochLength,
-  epochStability: EpochStability,
+  securityParameter: SecurityParameter,
   activeSlotCoefficient: ActiveSlotCoefficient,
   decentralizationLevel: DecentralizationLevel,
   desiredPoolNumber: number,
@@ -115,7 +115,7 @@ export type GetNetworkParametersApiResponse = {
   blockchain_start_time: string,
   slot_length: SlotLength,
   epoch_length: EpochLength,
-  epoch_stability: EpochStability,
+  security_parameter: SecurityParameter,
   active_slot_coefficient: ActiveSlotCoefficient,
   decentralization_level: DecentralizationLevel,
   desired_pool_number: number,